feat: Filter constants by visibility
- Visibility filters will also be applied to constants
